Add basis for further enhancing of function/method calls
