+++ WLC/code_executor/executor.py
@@ -6,7 +6,7 @@ import docker
 
 from WLC.code_executor.executor_error import ExecutorError, PY_STR_ERR_SYNTAX
 from WLC.code_executor.redirected_std import redirected_std
-from WLC.code_fixing.trial_codefixer import TrialCodeFixer
+from WLC.code_fixing.codefixer import TrialCodeFixer
 from WLC.image_processing.preprocessor import Preprocessor
 
 LOGGER = logging.getLogger()
