+++ WLC/code_fixing/codefixer.py
@@ -1,183 +1,388 @@
-import keyword
 import logging
-import string
+import sys
+from math import ceil
 
-from fuzzysearch import find_near_matches
+import editdistance
+import regex
+from stdlib_list import stdlib_list
 
-KW_LIST = keyword.kwlist + ["print", "list", "dict", "set", "file", "open", "assert", "range"]
+from WLC.code_fixing.static import get_functions
+
+PERMUTATION_LENGTH = 2
+ALLOWED_DIFFERENCE = 0.2
+
 LOGGER = logging.getLogger()
 
-SPEC_ATTR = ["doc", "name", "qualname", "module", "defaults", "code", "globals", "dict", "closure",
-             "annotations", "kwdefaults", "main"]
 
-
-class CodeFixer:
+class TrialCodeFixer:
     def __init__(self, code, indents, poss_lines):
         self.code = code
         self.indents = indents
         self.poss_lines = poss_lines
-        self.context = []
+        self.context = {'variables': [], 'functions': get_functions(), 'classes': [], 'imports': [], 'methods': []}
+        self.in_class = False
+        self.class_indent = 0
+        self.syntax = []
+        self.rules = []
+        self.statements = []
+
+        self.syntax.append(('VARIABLE', '[a-z_]\w*'))
+        self.syntax.append(('FUNCTION', '[a-z_]\w*'))
+
+        self.syntax.append(('STATEMENT', '.*'))
+        self.syntax.append(('BOOLEAN', '.*'))
+        self.syntax.append(('PARAMETERS', '.*'))
+
+        self.statements.append(('(FUNCTION)\((STATEMENT)\)', 2, None, self.fix_function_call))
+        self.statements.append(('(VARIABLE)\.(FUNCTION)\((STATEMENT)\)', 3, None, self.fix_method_call))
+        self.statements.append(('(VARIABLE)', 0, None, self.fix_variable))
+
+        # self.rules: list of quad-tuples (string to match, number of fixed, analysis_func, fix_func)
+        # analysis -> goes over result and gets any context var
+        # fix_func -> fixes the str with context var
+        self.rules.append(('import (.*)', 7, None, self.fix_import))
+        self.rules.append(('import (.*?) as (.*)', 11, None, self.fix_import_as))
+        self.rules.append(('from (.*?) import (.*?)(, .*?)*', 13, None, self.fix_from_import))
+        self.rules.append(('def (VARIABLE)\((PARAMETERS)\):', 7, self.analyze_def, self.fix_def))
+        self.rules.append(('class (VARIABLE):', 7, self.analyze_class, self.fix_class))
+        self.rules.append(('if (BOOLEAN):', 4, None, self.fix_if))
+        self.rules.append(('elif (BOOLEAN):', 6, None, self.fix_elif))
+        self.rules.append(('return (STATEMENT)', 7, None, self.fix_return))
+        self.rules.append(('while (BOOLEAN):', 7, None, self.fix_while))
+        self.rules.append(('for (VARIABLE) in (STATEMENT):', 9, self.analyze_for, self.fix_for))
+        self.rules.append(('for (VARIABLE) in range\((STATEMENT)\):', 16, self.analyze_for_range, self.fix_for_range))
+        self.rules.append(('(FUNCTION)\((STATEMENT)\)', 2, None, self.fix_function_call))
+        self.rules.append(('(VARIABLE)\.(FUNCTION)\((STATEMENT)\)', 3, None, self.fix_method_call))
+        self.rules.append(('(VARIABLE) = (STATEMENT)', 3, self.analyze_assignment, self.fix_assignment))
+        self.rules.append(('assert (STATEMENT)', 7, None, self.fix_assert))
+        self.rules.append(('del (STATEMENT)', 4, None, self.fix_del))
+        self.rules.append(('raise (STATEMENT)', 6, None, self.fix_raise))
+        self.rules.append(('pass', 4, None, lambda x, y: 'pass'))
+        self.rules.append(('else:', 5, None, lambda x, y: 'else:'))
+        self.rules.append(('break', 5, None, lambda x, y: 'break'))
+        self.rules.append(('continue', 8, None, lambda x, y: 'continue'))
+        self.rules.append(('(.*)', 0, None, None))  # If nothing else works this will
 
     def fix(self):
+        """
+        Main function to be called which finds the closest regex match for each line, extracts context variables and
+        function names and then attempts to fix typos.
+
+        :return: Fixed version of the code
+        """
+        LOGGER.debug('Starting code fixing.')
+
         fixed_lines = []
-        lines = self.code.splitlines()
+        closest_matches = []
 
-        # Iterate over all "root" elements to create scope, not actually changing the lines yet
-        for idx, (indent, line) in enumerate(zip(self.indents, lines)):
-            if indent == 0:
-                self._fix_line(line, idx)
+        LOGGER.debug('Compiling main rules.')
+        regexes = self.compile_regexes(self.rules)
 
-        for idx, line in enumerate(lines):
-            new_line = self._fix_line(line, idx)
-            fixed_lines.append(new_line)
+        LOGGER.debug('Looking for closest matches.')
+        for i in range(len(self.poss_lines)):  # range loop OK because of indexing type
+            closest_match = self.find_closest_match(self.poss_lines[i], regexes)
+            (match, analyze_func, _) = closest_match
+
+            if analyze_func:
+                analyze_func(match.groups(), i)
+
+            # At each line, check if currently in a class declaration.
+            if self.indents[i] < self.class_indent and self.in_class:
+                self.in_class = False
+
+            closest_matches.append(closest_match)
+
+        LOGGER.debug('Fixing lines.')
+        for idx, closest_match in enumerate(closest_matches):
+            (match, _, fix_func) = closest_match
+            if fix_func:
+                fixed_lines.append(fix_func(match, self.poss_lines[idx]))
+            else:
+                groups = match.groups()
+                LOGGER.debug("No match found for {}. Defaulting to not fixing.".format(*groups[1:]))
+                fixed_lines.append('{}'.format(*groups[1:]))
 
         return "\n".join("{indent}{code}".format(indent="  " * indent, code=line) for indent, line in
                          zip(self.indents, fixed_lines))
 
-    def _levenshtein(self, word, check_on=None):
-        if not check_on:
-            check_on = self.context + KW_LIST
+    def find_closest_match(self, poss_line, regexes):
+        """
+        Finds the closest regex to the line provided looking at possible permutations of characters.
 
-        start = 0  # doesn't change
-        end = 0
-        l_dist = 10  # max change is 2 anyway
-        curr_best_word = None
+        :param poss_line: Possible characters that make up the line. Type: [[Char]]
+        :param regexes: Compiled list of regexes
+        :return: The closest match and the functions used to analyze and fix the line
+        """
+        distance = sys.maxsize
+        closest = None
+        permutations = self.permutations(poss_line)
 
-        for kw in check_on:
-            pos = find_near_matches(kw, word, max_l_dist=min(2, len(kw) - 1), max_insertions=0, max_deletions=0)
+        for r, fixed, analyze, fix in regexes:
+            for possible in permutations:
+                match = r.match(possible)
 
-            for p in pos or []:
-                if p.start == start:
-                    # If word is longer, or if theres less l_dist (prioritising contextual_data)
-                    if end < p.end or (end == p.end and p.dist < l_dist and not (
-                                    curr_best_word in self.context and kw in KW_LIST)):
-                        end = p.end
-                        curr_best_word = kw
-                        l_dist = p.dist
+                if match:
+                    # use fuzzy counts here
+                    if sum(match.fuzzy_counts) - fixed < distance:
+                        distance = sum(match.fuzzy_counts) - fixed
+                        closest = (match, analyze, fix)
 
-        return curr_best_word, start, end
+                        if distance < 0:
+                            return closest
 
-    def _fix_word(self, word, word_number=None, line_number=None, prev_word="", next_word=""):
-        if prev_word == "import" or prev_word == "for" or next_word == "=":
-            if word and word not in self.context:
-                self.context.append(word)  # add the imported module, or the var name to contextual data
-            return word
+        return closest
 
-        elif word.startswith('--') and word.endswith('--'):
-            no_dash = word.strip('--')
-            best_word, start, end = self._levenshtein(no_dash, check_on=SPEC_ATTR)
-            if best_word and len(best_word) == len(no_dash):
-                word = "__" + best_word + "__"
-            else:
-                word = "__" + no_dash + "__"
+    def permutations(self, poss_chars):
+        """
+        Finds the most probable permutations of the line. Complexity is reduced by the fact that there are only 1 or 2
+        possibilities of a character.
 
-        elif not prev_word:
-            best_word, start, end = self._levenshtein(word)
-            if best_word:
-                word = word[:start] + best_word + word[end:]
+        :param poss_chars: Possible characters that make up the line
+        :return: Possible ways to write the line returned as a list of strings
+        """
+        if not poss_chars:
+            return ['']
 
-        return word
+        results = []
+        permutations = self.permutations(poss_chars[1:])
 
-    def _fix_line(self, line, line_number):
-        fixed_words = []
-        words = line.split()
-        for word_number, word in enumerate(words):
-            prev_word = fixed_words[-1] if fixed_words and fixed_words[-1] in ["def", "class", "import"] else None
-            next_word = words[word_number + 1] if len(words) > word_number + 1 else None
+        for char in poss_chars[0][:PERMUTATION_LENGTH]:
+            for permutation in permutations:
+                results.append(char + permutation)
 
-            fixed_word = word  # set default
+        LOGGER.debug('Trying %s permutations', len(permutations))
 
-            if prev_word == "class" or prev_word == "def":  # class/method/function declaration
-                self._set_decl(word)
+        return results
 
-            elif prev_word == "for":  # se var X as context in a 'for X in _' statement
-                self._set_forloop_var(word)
+    def compile_regexes(self, rules):
+        """
+        Compiles the list of regexes replacing keywords in the rules by the syntax expressions. Syntax expressions can
+        also refer to other expressions that were listed before them because they get expanded.
 
-            # TODO: check valid function name
-            elif "(" in word and ")" in word and len(word[:word.find('(')]) > 0:  # method/function call + prefixed by function name
-                fixed_word = self._fix_funccall(word, line_number, word_number)
+        :param rules: List of rules that should get compiled into syntax expressions
+        :return: Compiled list of regexes
+        """
+        regexes = []
 
-            # TODO: functions with spaces in args
-            else:  # not detected as anything special - attempt to autocorrect
-                fixed_word = self._fix_word(word, word_number, line_number,
-                                            prev_word=prev_word, next_word=next_word)
+        for i in range(len(self.syntax)):
+            for j in range(i):
+                self.syntax[i] = (self.syntax[i][0], self.syntax[i][1].replace(self.syntax[j][0], self.syntax[j][1]))
 
-            if prev_word == "=" and len(fixed_words) > 2:  # shouldve already been caught but sanity check
-                self._set_var(fixed_words)
+        for rule, fixed, analyze, fix in rules:
+            difference = max(ceil(fixed * ALLOWED_DIFFERENCE), 0)
+            reg = '^(?e)((?:%s){e<=%s})$' % (rule, difference)
 
-            if word != fixed_word:
-                LOGGER.debug("Replaced %s with %s", word, fixed_word)
-            fixed_words.append(fixed_word)
+            for i in range(len(self.syntax)):
+                reg = reg.replace('({})'.format(self.syntax[i][0]), '({})'.format(self.syntax[i][1]))
 
-        return " ".join(fixed_words)
+            r = regex.compile(reg)
+            regexes.append((r, difference, analyze, fix))
 
-    def _fix_funccall(self, word, line_number, word_number):
-        # TODO: add support for multiple args -> requires more context due to spaces, commas etc
-        argstart = word.find("(") + 1
-        argend = word.rfind(")")
-        args = word[argstart:argend]
+        regexes = sorted(regexes, key=lambda x: -x[1])
 
-        # attempt to fix the called func_name being invalid
-        func_name = word[:argstart - 1]
-        if func_name not in self.context:
-            fixed_func_name = self._fix_word(func_name)
-            if fixed_func_name in self.context:
-                word = func_name + word[argstart - 1:]
+        return regexes
 
-        if args in self.context or all(
-                        arg_char in string.digits for arg_char in args):  # TODO: add support for string args
-            pass  # all good here, valid variable or int
+    def levenshtein_closest(self, poss_chars, possibilities, allowed_difference=ALLOWED_DIFFERENCE):
+        """
+        Finds one of the possible strings which is closest to a permutation of a line or part of a line. This is fairly
+        expensive when there are a lot of possibilities and permutations.
 
-        elif args:
-            fixed_args = self._fix_args(args, word, line_number, word_number)
-            word = word[:argstart] + fixed_args + word[argend:]
+        :param poss_chars: Possible characters to make up all of the permutations of the line/substring
+        :param possibilities: Possible strings which should be matched to the closest permutation
+        :param allowed_difference: Maximum percentage difference between closest match and one possibility
+        :return: The closest possibility to the provided line/substring and the levenshtein distance
+        """
 
-        return word
+        permutations = self.permutations(poss_chars)
+        best = allowed_difference
+        recommended = permutations[0]
 
-    def _fix_args(self, args, fixed_word, line_number, word_number):
-        new_args = args
-        some_args, start, end = self._levenshtein(args)
-        # TODO: multiple args/words
-        if "(" in args and ")" in args and len(args[:args.find('(')]) > 0: # is a nested function
-            new_args = self._fix_funccall(args, line_number, word_number)
+        for permutation in permutations:
+            for possibility in possibilities:
+                distance = editdistance.eval(permutation, possibility) / len(possibility)
 
-        elif some_args and some_args in self.context:
-            new_args = some_args
+                if distance == 0:
+                    return possibility, 0
+                elif distance < best:
+                    recommended = possibility
+                    best = distance
+        return recommended, best
 
-        elif args.startswith("--") and args.endswith("--"):
-            new_args = self._fix_word(args)
+    # ANALYSE
+    def analyze_class(self, groups, line_n):
+        LOGGER.debug("Analysing class. Adding {} to context.".format(groups[1]))
+        self.context['classes'].append(groups[1])
+        self.in_class = True
+        self.class_indent = self.indents[line_n]
 
+    def analyze_for_range(self, groups, line_n):
+        LOGGER.debug("Analysing range. Adding {} to context.".format(groups[1]))
+        self.context['variables'].append(groups[1])
+
+    def analyze_for(self, groups, line_n):
+        LOGGER.debug("Analysing for. Adding {} to context.".format(groups[1]))
+        self.context['variables'].append(groups[1])
+
+    def analyze_assignment(self, groups, line_n):
+        LOGGER.debug("Analysing assignment. Adding {} to context.".format(groups[1]))
+        self.context['variables'].append(groups[1])
+
+    def analyze_def(self, groups, line_n):
+        LOGGER.debug("Analysing function def. Adding {} to context.".format(groups[1]))
+        # differentiate between functions and class methods
+        if self.indents[line_n] <= self.class_indent and self.in_class:
+            self.context['methods'].append(groups[1])
         else:
-            # we should use digits
-            possibles = self.poss_lines[line_number][word_number]
-            poss_args = ""
-            start_pos = fixed_word.find(args)
-            curr_possibilities = {i: possibles[i] for i in range(start_pos, start_pos + len(args))}
-            for char, possible_char in zip(args, curr_possibilities.values()):
-                top3 = [digit_chars for digit_chars in possible_char[1:3] if str(digit_chars) in string.digits]
-                if any(top3):  # if any of top3 are digits
-                    poss_args += top3[0]
-            if len(poss_args) == len(args):
-                new_args = poss_args
+            self.context['functions'].append(groups[1])
+        variables = groups[2].split(',')
+        LOGGER.debug("Analysing function def variables. Adding {} to context.".format(variables))
+        self.context['variables'].extend(variables)
 
-        return new_args
+    def fix_import(self, match, poss_chars):
+        groups = match.groups()
+        poss_import = poss_chars[match.start(2): match.end(2)]
 
-    def _set_decl(self, fixed_word):
-        if "(" in fixed_word:
-            name = fixed_word[:fixed_word.find("(")]
-            if name and name not in self.context:
-                self.context.append(name)
+        # poss_import = self.extract_poss_chars(groups[0], poss_chars, groups[1])
+        closest, _ = self.levenshtein_closest(poss_import, stdlib_list("3.6"))
+        LOGGER.debug("Fixing import. Adding {} to context after analysis.".format(closest))
+        self.context["imports"].append(closest)
+        return 'import {}'.format(closest)
 
-        if "(" in fixed_word and ")" in fixed_word:
-            args = fixed_word[fixed_word.find("(") + 1:fixed_word.rfind(")")]
-            if args and args not in self.context:
-                self.context.append(args)
+    def fix_import_as(self, match, poss_chars):
+        groups = match.groups()
+        poss_import = poss_chars[match.start(2): match.end(2)]
+        closest_module, _ = self.levenshtein_closest(poss_import, stdlib_list("3.6"))
+        LOGGER.debug("Fixing import as. Adding {} to context after analysis.".format(groups[2]))
+        self.context["imports"].append(groups[2])
+        return 'import {} as {}'.format(closest_module, groups[2])
 
-    def _set_var(self, fixed_words):
-        var_name = fixed_words[-2]
-        if var_name and var_name not in self.context:
-            self.context.append(var_name)  # add var name before = sign to contextual data
+    def fix_from_import(self, match, poss_chars):
+        groups = match.groups()
+        poss_import = poss_chars[match.start(2): match.end(2)]
+        # poss_import = self.extract_poss_chars(groups[0], poss_chars, groups[1])
+        closest_module, _ = self.levenshtein_closest(poss_import, stdlib_list("3.6"))
+        LOGGER.debug("Fixing from X import Y. Adding {} to context after analysis.".format(closest_module))
+        for imported in groups[1:]:
+            self.context["imports"].append(imported)
+        return 'from {} import {}'.format(closest_module, ", ".join(groups[1:]))
 
-    def _set_forloop_var(self, var):
-        if var and var not in self.context:
-            self.context.append(var)
+    def fix_class(self, match, poss_chars):
+        groups = match.groups()
+        poss_class = poss_chars[match.start(2): match.end(2)]
+        # poss_class = self.extract_poss_chars(groups[0], poss_chars, groups[1])
+        closest, _ = self.levenshtein_closest(poss_class, self.context["classes"])
+        LOGGER.debug("Fixing class. Changing from {} to {}.".format(groups[1], closest))
+        return 'class {}:'.format(closest)
+
+    def fix_if(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing if. Using {}.".format(*groups[1:]))
+        return 'if {}:'.format(*groups[1:])
+
+    def fix_elif(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing elif. Using {}.".format(*groups[1:]))
+        return 'elif {}:'.format(*groups[1:])
+
+    def fix_return(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing return. Using {}.".format(*groups[1:]))
+        return 'return {}'.format(*groups[1:])
+
+    def fix_while(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing while. Using {}.".format(*groups[1:]))
+        return 'while {}:'.format(*groups[1:])
+
+    def fix_for(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing for. Using {} and {}.".format(*groups[1:]))
+        return 'for {} in {}:'.format(*groups[1:])
+
+    def fix_for_range(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing for range. Using {} and {}.".format(*groups[1:]))
+        return 'for {} in range({}):'.format(*groups[1:])
+
+    def fix_function_call(self, match, poss_chars):
+        groups = match.groups()
+        poss_func = poss_chars[match.start(2): match.end(2)]
+        # poss_func = self.extract_poss_chars(groups[0], poss_chars, groups[1])
+        closest, _ = self.levenshtein_closest(poss_func, self.context["functions"])
+        LOGGER.debug(groups)
+        LOGGER.debug("Fixing func call. Using {} and {}.".format(*(closest, *groups[2:])))
+
+        # use 3 not 2 because of list 0-index
+        new_args = self.fix_arguments(groups[2], poss_chars[match.start(3): match.end(3)])
+
+        LOGGER.debug("Fixing func call arguments from {} to {}".format(groups[2], new_args))
+        return '{}({})'.format(closest, new_args)
+
+    def fix_method_call(self, match, poss_chars):
+        groups = match.groups()
+        poss_var = poss_chars[match.start(2): match.end(2)]
+        # poss_var = self.extract_poss_chars(groups[0], poss_chars, groups[1])
+        closest_var, _ = self.levenshtein_closest(poss_var, self.context["variables"])
+        LOGGER.debug("Fixing method call var. From {} to {}".format(groups[1], closest_var))
+
+        poss_method = poss_chars[match.start(3): match.end(3)]
+        # poss_method = self.extract_poss_chars(groups[0], poss_chars, groups[2])
+        closest_method, _ = self.levenshtein_closest(poss_method, self.context["methods"])
+        LOGGER.debug("Fixing method call method. From {} to {}".format(groups[2], closest_method))
+
+        closests_args = [
+            self.levenshtein_closest(poss_arg, self.context["variables"])[0] for poss_arg
+            in (poss_chars[match.start(group_n): match.end(group_n)] for group_n in range(4, len(groups)+1))
+        ]
+        LOGGER.debug("Fixing method call args. From {} to {}.".format(groups[3], closests_args))
+
+        return '{}.{}({})'.format(closest_var, closest_method, *closests_args)
+
+    def fix_assignment(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing assignment. Using {} = {}.".format(*groups[1:]))
+        return '{} = {}'.format(*groups[1:])
+
+    def fix_assert(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing assert. Using {}.".format(*groups[1:]))
+        return 'assert {}'.format(*groups[1:])
+
+    def fix_del(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing del. Using {}.".format(*groups[1:]))
+        return 'del {}'.format(*groups[1:])
+
+    def fix_raise(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing raise. Using {}.".format(*groups[1:]))
+        return 'raise {}'.format(*groups[1:])
+
+    def fix_def(self, match, poss_chars):
+        groups = match.groups()
+        LOGGER.debug("Fixing def. Using {} and {}.".format(*groups[1:]))
+        return 'def {}({}):'.format(*groups[1:])
+
+    def fix_arguments(self, args, poss_chars):
+        # TODO: THIS DOESNT DO BRACKET MATCHING PROPERLY -> two functions in one decl will break due to greedy matching
+        # need to write a simple parser to split arguments while keeping position and matching brackets.
+        arg_matches = regex.finditer("[ ]?([^,()]+(?:\(.*\))?)", args)
+
+        # TODO: some statement matching here - func vs variable vs random
+        result = []
+        for match in arg_matches:
+            group = match.groups()[0]
+            # NOTE: commented top avoid confusion.
+            # new_arg = self.fix_variable(match, poss_chars[match.start(1): match.end(1)])  # when variable.
+            new_arg = group
+            LOGGER.debug("Fixing argument {} to {}".format(group, new_arg))
+            result.append(new_arg)
+
+        return ", ".join(result)
+
+    def fix_variable(self, match, poss_chars):
+        groups = match.groups()
+
+        closest, _ = self.levenshtein_closest(poss_chars, self.context["variables"])
+        LOGGER.debug("Fixing variable {} to {}.".format(groups[0], closest))
+        return closest
